Modification in the regex of the startWithNumber function in XGetter to include those objects that include language in the quality field. It would be interesting to include a language field in the Xmodel. (for example to store the language of those servers like UptoStream that give this information)
Modification of the SortMe function in XGetter, to include those objects (links) when the quality attribute is empty (for example Openload server)

Modification in the regex of the startWithNumber function in XModel to include those objects that include language in the quality field. It would be interesting to include a language field in the Xmodel. (for example to store the language of those servers like UptoStream that give this information)